Initial Check-in of ECMPS Check Engine compiled using .Net Core 3.1.
Contains Windows references in the DatabaseAccess and TestingPlatform projects.  The DatabaseAccess reference exists because of SQL Server use.  The TestingPlatform exist because of the WinForm it contains, which is not a complete form.

THIS VERSION HAS NOT BEEN TESTED.  It just compiles.

I did run all of the MS Unit Tests and corrected all that failed except for three.  The three need to be re-thought and rewritten.